feat(cli): add rollups create command
