+++ .changeset/fresh-coats-know.md
@@ -1,0 +1,5 @@
+---
+"@cartesi/cli": major
+---
+
+DEPRECATED: cartesi create
