+++ .changeset/poor-tips-happen.md
@@ -1,0 +1,5 @@
+---
+"@cartesi/cli": major
+---
+
+new: cartesi rollups create
